chore: Move Stack to internal/stack
